+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/setting/SettingViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.boostcampwm2023.snappoint.data.repository.RoomRepository
 import com.boostcampwm2023.snappoint.data.repository.SignInRepository
-import com.boostcampwm2023.snappoint.presentation.util.SignInUtil
+import com.boostcampwm2023.snappoint.presentation.util.UserInfoPreference
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -18,7 +18,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SettingViewModel @Inject constructor(
-    private val signInUtil: SignInUtil,
+    private val userInfoPreference: UserInfoPreference,
     private val signInRepository: SignInRepository,
     private val roomRepository: RoomRepository
 ) : ViewModel() {
@@ -32,7 +32,7 @@ class SettingViewModel @Inject constructor(
     fun onSignOutClick() {
         signInRepository.getSignOut()
             .onEach {
-                signInUtil.clearUserAuthData()
+                userInfoPreference.clearUserAuthData()
                 _event.emit(SettingEvent.SignOut)
             }
             .catch {
@@ -46,7 +46,7 @@ class SettingViewModel @Inject constructor(
     }
 
     fun getSavedPost() {
-        roomRepository.getAllLocalPost(signInUtil.getEmail())
+        roomRepository.getAllLocalPost(userInfoPreference.getEmail())
             .onEach {
                 Log.d("LOG", it.toString())
             }.catch {
